Fix attempts tool and add level background

+++ Model/BlockData.cs
@@ -7,7 +7,7 @@ namespace test_raylibs.Model
     class BlockData
     {
         public string type { get; set; }
-        public float x { get; set; }
-        public float y { get; set; }
+        public int x { get; set; }
+        public int y { get; set; }
     }
 }
